update: use the new dentry walk logic
